+++ lib.js
@@ -139,16 +139,24 @@ export function getClient() {
  * @param {string} path Path to the proof file.
  */
 export async function readProof(path) {
+  let bytes
   try {
-    await fs.promises.access(path, fs.constants.R_OK)
+    const buff = await fs.promises.readFile(path)
+    bytes = new Uint8Array(buff.buffer)
   } catch (/** @type {any} */ err) {
     console.error(`Error: failed to read proof: ${err.message}`)
     process.exit(1)
   }
+  return readProofFromBytes(bytes)
+}
 
+/**
+ * @param {Uint8Array} bytes Path to the proof file.
+ */
+export async function readProofFromBytes(bytes) {
   const blocks = []
   try {
-    const reader = await CarReader.fromIterable(fs.createReadStream(path))
+    const reader = await CarReader.fromBytes(bytes)
     for await (const block of reader.blocks()) {
       blocks.push(block)
     }
@@ -156,7 +164,6 @@ export async function readProof(path) {
     console.error(`Error: failed to parse proof: ${err.message}`)
     process.exit(1)
   }
-
   try {
     // @ts-expect-error
     return importDAG(blocks)